Rename onboarding_func to onboarding and delete class version
